backend/requirements.txt
# Most Recent Version Packages
setuptools
wheel
djangorestframework
django-cors-headers